Remove --skip-build option from `release create`
This is getting in the way / making refactoring to enable upstreams really
complicated.

This was designed originally to allow building the installers without having to
rebuild everything. Rebuilding everything if you already have a build locally is
now really fast. Skipping all the build there is the `create_installer`. Once
upstreams are done, pkgpanda will likely be able to just download the last build
of the relevante packages from CI