Do a better job fixing cached bootstrap download
The work dir should be the packages directory so relative path file fetches work how a user might expect.

The output filename is in the cache directory.

PR #93 fixed part of the problem but not all of it, as can be seen in the build:  https://teamcity.mesosphere.io/viewLog.html?buildId=260177&tab=buildResultsDiv&buildTypeId=DcosIo_Dcos_Master