Allow to delete Exhibitor S3 bucket after the fact
By adding its resource ID to the outputs of the CloudFormation stack.

The bucket is created with a "Retain" deletion policy, which means that
destroying the stack won't delete the bucket. This change makes it
possible to still delete the bucket in an automated way (figuring out
the bucket name without the stack output is much more complicated).