if CI_FLAGS is set, exit with code 0, since the CI should pick up the results and fail that way
Exiting non-zero causes TeamCity to hard fail, even if the test which failed is muted.

This makes CI_FLAGS interpreted as a sign that we're running in a CI environment which understands
structured test results, and that the CI system will interpret failed test as a failed run.

So the new method is either "basic" environment where a non-zero exit indicates error, or
"everything ran successfully but some tests failed" indicated by test results being reported out and
the process exiting zero.

Fixes:
 - Make exception catching more precise