Move to dynamic ports for Minuteman and Octarine test apps
The marathon ports API is weird, when you don't say  requirePors, the port being specified is the apps "servicePort", and the app gets launched / assigned a random port on the host. The tests ended up assigning multiple apps the same service port (Which gets caught by newer versions of marathon).

This switches all of the ports to always be dynamic from Marathon so we don't have to keep track of any interference. Another option would have been to use `requirePorts` and change the ports to be unique by hand.