Teach waiting for exhibitor to "shortcut" if the local zookeeper is known to be running with a good config
Previously whenever a service, such as a mesos master restarted, it would have
to wait until all zookeeper instances in the cluster were started and stable.
This meant that if one master was gone, gone, no other master could be restarted
(it would never come back up until the gone node was replaced).

This changes that logic a bit, making it so that if the exact same zookeeper
process which was running when the mesos master (or friends) comes up is still
running then the master can start up instantly. This should be safe beause that
zookeeper had to have been started with a full quorum config by exhibitor, which
means it should be up, running, with a valid state of the cluster overall,
rather than this node belonging to a subset / partition which hasn't reached
overall cluster status.