Include localhost entry on Azure clusters
Fixes an issue currently impacting Azure regions in Australia and North
Europe, where the Azure DNS servers are no longer responding with an
entry for `localhost`. The startup service currently overwrites the
default /etc/hosts file so that it is an idempotent operation and can be
run multiple times if the $(detect_ip) value changes.