Fix AWS template validation
AWS templates were not being validated even when the release
configuration contains the necessary credentials. This was because the
get_session function was being called with the top-level config object,
but it should be receiving the ['testing']['aws'] values.

Enable testing by adding testing section for AWS to the default release
configuration.