Enable the Marathon 'gpu_resources' feature by default.
By default, enabling this feature will have absolutely no affect on the
existing operation of marathon.

However, it makes it much easier for people to use GPUs in DC/OS if they
wish to go through the manual process of enabling the mesos GPU isolator
on a per-agent basis via /opt/mesosphere/etc/mesos-slave-common.

In future releases, we should come up with a way to make enabling this
isolator less manual, but for now enabling this marathon feature by
default should suffice for early users.