Replace isort with flake8-import-order
This replaces isort which is painful to use / maintain with the
flake8-import-order plugin which means we don't have do things like manually
maintain a whitelist for isort and a blacklist for flake8. Don't have to remember
to add things to isort. Don't have to manually set known_first_party and known_third_party

Style wise the biggest difference from current the "smarkets" style which is set
now is that from `x import a, B, c` cares about sorting alphabetically without
caring about case now

Based on the code changes, this catches a bunch of basic sorting was off which says the tool config
for isort was too complicated and the new plugin is a lot easier in that respect.

I think the improvement in CI usability (They show up on the teamcity base page) and catching
programmer mistakes / harder to forget to add things to be isorted, this is worth the slightly worse
diagnostics in telling people how to fix it (isort would print a diff and auto-fix, this just says "higher or lower")