Rework gen/ to be made of three primitives: Source, Target, Scope and Resolver
Source is a source of configuration variables. A target is a set of configuration variables which
need to be set to have a complete configuration. The resolver figures out how to use a collection of
sources to fill in a collection of targets and check for conflicts.

Sources can be user-provided (a dictionary loaded out of a config file). They can also be provided
by different contexts (AWS templates have a source of AWS which implies things likes resolvers). A
third common provider is deriving from other already given values (master count is derived from a
master list if a master list is given).

Targets are a set of variables which need to be set, as well as a set of variables which need to be
set under certain conditions. The "conditional" variables you can think of as a "Switch". Based on
the value of the argument to the switch, one of multiple paths are going to be taken, and all
configuration variables in that particular path are expected to be taken.

The Resolver takes a collection of sources and targets, then tries to calculate (and validate) all
the values needed to satisfy the targets. It keeps track of the path to get to a specific value, if
there are any conflicts between different sources (master_count is set to a conflicting / different
value than the length of master_list).