Revert "Make dcos-spartan-watchdog start later rather than have a random 60 second sleep"
This reverts commit ecfe9a57ac9e61e84ad84161c8854f51aa12f417.

There still isn't a good explanation _why_ sleep is better than a delay. The behavior hasn't
really been explained.

That said, this fixes the issue for one customer at present, although we still need to come up with
a reasonable reason for that and _proper_ fixes. `sleep` is rarely the correct solution to software
problems, just a band-aid...