Move to using Sources rather than manual argument dictionary construction
This removes the `providers` switch in gen/calc.py, pushing all the data which
was somewhat hackishly combined there down into the correct installer components
using the "Source" concept of gen to provide the values. This means moving onprem,
AWS, and Azure to use "Source" directly and add the source to the set of sources
when calling gen.generate rather than setting a magic key by hand and relying
on that to "imply" the right configuration.

This should provide better error messages for all cases: There are less bad
paths which we can tell the user about for possibly setting configuration. In
addition, it makes lite binding configuration easier to implement, as all the
pieces start being more "bound" to the one particular installer they effect.

From my point of view this also makes the code easier to reason about than the
previous nested conditionals which selected which options to use.

This also moves the AWS Simple, AWS Advanced,  and Azure templates to all use
lists of sources to manage their argument addition rather than manipulating a
big dictionary of arugments by hand, which makes it considerably harder to do
things like accidentally unconditionally overwrite user configuration resulting
in oddness at runtime.

The num_masters Sources are also unified to all operate / be constructed the
same way for all the templates.

Finally, the list of sources is brought in as part of the "config id" so that
when two templates have different sources (say aws single-master simple and aws
multi-master simple template), then they will end up with different config ids.
Originally they were ending up with the same config id which meant the one built
last's config would end up in both's cluster, which resulted in the
single-master tmeplates not being launchable