Update / rewrite ssh.Tunnel, TunnelCollection
 - TunnelCollection is replaced with a contextlib.ExitStack, so it has less code.
 - Tunnel class is replaced with a "contextmanager" function that yields a
   Tunneled object, which provides the same interface that `with Tunnel` did.

The tunnel contextmanager function makes it harder to use the tunnels incorrectly
as the tunnel is only accessible inside the contextmanager / when the
contextmanager has been entered. The "Tunneled" class only has the simple API /
behavior people expect to work on things which are tunneled to.

Moving TunnelCollection to be the tunnel_collection function makes it so we're
no longer having to access `tunnels.tunnels` in order to get the tunnels.