Replace type assertions with python 3.5 type annotations
The set of things to replace was found with:
`git grep 'assert isinstance'`

The python 3.5 type hints are now expressive enough for the all of these,
through the new "typing" module. The hints will effectively do the same thing
the asserts used to.

Some type assertions still exist, particularly in cases where there are circular
references (referring to a not-defined-yet class), as well as inside gen
validation where we're explicitly trying to walk through the type.