Remove magic global passing of azure and aws storage urls
Previously these were passed by globals when we were doing a release and as arguments when we did a advanced / custom configuration template generation. Now they are always passed as arguments which are injected by release/__init__.py in the case of release for the configuration, and via the normal argumen means otherwise.

The magic config global still exists so aws CF template validation can work, although that can definitely have the same pattern done to it in the future. It's more complicated to do however since it needs to ensure the AWS keys that are passed through don't end up inside the produced cluster