Modified `mesos-modules` to build against a specific branch.
Currently the `mesos-modules` package in DC/OS is built against the
`dcos/dcos-mesos-modules` master. The Mesos modules in the
`dcos/dcos-mesos-modules` repo has dependencies on Mesos headers and should
always be compiled against the Mesos source include in DC/OS
(mesosphere/mesos). Hence, we have introduced branches in the
`dcos/dcos-mesos-modules` to track the version of DC/OS. This would help
maintainability of the `dcos/dcos-mesos-modules` repo when we want to fix
issues in the Mesos modules against minor versions of DC/OS.

Since, we have branches in the `dcos/dcos-mesos-modules` repo we should be
compiling the `mesos-modules` package against a specific branch that conforms
to the DC/OS Mesos version.