Default oauth_available to true in base source
We aren't yet able to override config values in a downstream variant if
they've been defined outside the base source entry in gen/calc.py. We
need to define oauth_available since we can't derive it from
oauth_enabled, which is late-binding, but we also need to be able to
override security-related configuration downstream. Both needs are met
by setting this default.