Fix package list validation during build
This prevents an exception from being raised when a bootstrap package
list doesn't include a package listed in a treeinfo's package variants.
A treeinfo with bootstrap_package_list is expected to produce a
bootstrap tarball containing a subset of all cluster packages, so it's
not necessarily an error if a package variant isn't present in it.