Turned on the mesos agent flag '--disallow_sharing_agent_pid_namespace'.
This mesos agent flag is used for security concern by cluster operators.
In favor of MESOS-7853, we allow users to configure whether a container
can share the PID namespace from its parent. If the container is a top
level container (e.g., executor containers), it will share the PID
namespace from the agent. Operators might want to limit users to share
the agent's PID namespace by accident.

So we should turn on this flag in DC/OS. Every top level container in
DC/OS could not access to the agent's PID namespace. Note that turning
on this agent flag does not change any current semantic.