Remove LD_LIBRARY_PATH override for executors
This modification to the linker's path is confusing, and can result in non-DCOS
binaries being linked against DCOS libraries. This has caused breakage in at
least one instance where a system-provided `curl` linked against an interface-
incompatible older `libcurl` provided by DCOS, causing MESOS_HTTP healthchecks
to fail.

This commit mirrors the path started in dcos/dcos#707, but with the addition of
rpath in 27339e3 that change is no longer necessary. Further, we opt to preserve
the SASL_PATH override, even though it suffers similar concerns, because libsasl2
would require special mesos-only configuration at compilation time in order to
provide for mesos' out-of-the-box authentication, and that library is not
currently part of the mesos build flow.

Fixes DCOS_OSS-1302.