Remove port range from 14849 to 15103 from global port pool
Linux kernel 3.10.0-327 (centos 7.2) and 3.10.0-514 (centos 7.3) drops outgoing
lxlan tcp packets if the destination port is in the range from 14849 to 15103.
It's 255-port range, from 0x3a01 to 0x3aff. it's not iptables or something like
that, it's somewhere in interface driver. ICMP and UPD work fine. If you start
another tcp server outside the range 14849-15103 in the same container it will
work. We didn't find the exact line where this bug is happening in linux kernel,
but we haven't seen this bug on linux kernel 3.10.0-693 (centos 7.4).