bootstrap: More robust handling of a zk.pid file while waiting for exhibitor
There seems to be a way for the exhibitor to launch ZK in a way that
does not result in ZK PID file being created. Despite numerous attempts,
I was unable to recreate this situation but removing ZK PID *without*
restarting exhibitor service yielded the very same error log entries as
the ones reported by the issue submitter.

After a quick debug of the bootstrap.py code, it turns out that
exhibitor.py module used by bootstrap script makes a shortcut while
waiting for the local ZK by comparing the ZK PID file's mtime with the
one stored internally. The problem is though that the shortcut always
assumes that there is a PID file as long as there is a stashed mtime
value (see [1] -> [2] -> [3]). This was resulting in FileNotFound
exception being thrown as reported by the user.

This PR changes that:
* even if the mtime was stashed by the module, it is ignored by it if
the ZK PID is missing. In such case, the normal (i.e. long) wait for the
ZK quorum is performed and the "shortcut" is terminated early.
* after the normal wait for the ZK quorum has been performed, the module
does not try to stash the ZK PID's mtime if the PID does not exist.

This, of course, does not solve the underlying issue - i.e. why the ZK
did not create the PID or created it in the wrong place. But it will
allow the bootstrap script to handle the error more gracefully and in
cases when the ZK quorum, as seen by Exhibitor, is healthy, it will
allow the bootstrap script to proceed, even though the ZK PID is
missing. The extra logline that was added will maybe allow us to debug
the root cause at some point in the future.

[1] https://github.com/dcos/dcos/blob/98d0ac1f34737540a1f9b60fe9d4db2aeffc20db/packages/bootstrap/extra/dcos_internal_utils/exhibitor.py#L28-L33
[2] https://github.com/dcos/dcos/blob/98d0ac1f34737540a1f9b60fe9d4db2aeffc20db/packages/bootstrap/extra/dcos_internal_utils/exhibitor.py#L36
[3] https://github.com/dcos/dcos/blob/98d0ac1f34737540a1f9b60fe9d4db2aeffc20db/packages/bootstrap/extra/dcos_internal_utils/exhibitor.py#L20