Skip copying custom check bins package from cache
This fixes an error where, when custom check bins are provided, the
installer will attempt to copy the custom check bins package from the
Docker container's artifact cache, and fail with an error when it can't
find it. This is accomplished by adding the package to the gen result's
list of generated stable artifacts, which are skipped when copying
artifacts.