Added second EBS drive to agents and public agents.
Without this, agent nodes do not re-join the cluster after reboot.
This is due to our coreos-aws cloud-config where we expect a /dev/xvdb.

The time service is able to start before the cloud-config.yaml adds
the var-lib.mount unit. After reboot the problem shows up as the
units are on the disk at that point.