Add bootstrap ip-detect check for static backend
The ip-detect output should always match the expected master node IP
addresses in a static Exhibitor backend configuration of DC/OS. The
additional change in behavior for one master clusters is introduced to
keep things consistent. Without the change in behavior the static one
master setup with a diverging ip-detect script would not break. The fact
that every one master cluster gets a static backend but the IP is not
determined by the static backend setting in the configuration is
considered a bug.