Use total space available in mount disks when creating Mesos resources.
This concerns agents and the file `/var/lib/dcos/mesos-resources`.
We previously used the free space on the agent's mounted volumes
when using `make_disk_resources.py` to create the Mesos resources
file.

We recently saw that the Mesos agent would not restart due to this
behavior for a specific case: a mounted volume is used on an agent
by a framework, the agent is stopped and a second volume is added.
This is because, when adding an agent, we create again the Mesos
resources file and, if a mounted volume was used, its disk space
will have decreased. Mesos will then not accept to restart as it
cannot find the volume (due to the change of value).

To fix this, we now use the total amount of disk space available on
the volume. This will not cause problems when upgrading DC/OS as,
when `make_disk_resources.py` is first used, the free amount of
space on a mount volume is likely to be its total amount of space.