Python install for Windows, with associated pip and setuptools
Python 3.6.5 has pip and setuptools included but on Windows we
needed to install python to a temporary directory and copy to the
final directory as part of buiding. This broke pip and setuptools
and required the setups to be broken into two, one for core python,
and the second to bootstrap the install of pip and setup.