Fixes hard-coded python3.5 for local_build
This led to issues with newer Ubuntu versions that
ship with python3.6 (or anything newer than 3.5 in the future).
Ubuntu has the 'bug' that makes it a requirement to install pythons `venv`
module as a system package to be able to use `pip` inside any venv. See
https://bugs.launchpad.net/ubuntu/+source/python3.4/+bug/1290847.
This Ubuntu specific system package however only adds support for their latest
shipped python version, which (at least on 18.04, if not earlier) is 3.6.

Please note that all the workaround in the linked ticket do not work for our
use-case, as we have scripts that update pip inside the temporary venv.