adminrouter: remove user lookup from authn hot path
The fact that dcos-oauth is in the hot path of every
Admin Router authentication decision is known to violate
the DC/OS authentication architecture fundamentals.

Fundamentally, a specific type of JSON Web Token is to be
used as the DC/OS authentication token. Successful
cryptographic verification of a DC/OS authentication token
is neccessary and sufficient proof of authentication.
That is, authentication can generallay be done purely
locally w/o involving network communication. This has
great advantages with respect to reliability and
availability of the system (and some disadvantages
w.r.t. security, arguably). A subsequent database lookup
as performed by the current code undermines this concept.

Remove the database lookup (to ZooKeeper, through
dcos-oauth) from the hot path, and clean up some
other aspects of /auth/open.lua.