adminrouter: prepare for public key-based auth token verification
Prepare Admin Router's auth/open.lua code for the concept
of public key-based authentication token verification.
In that case the verification key is not a secret key
which is why this commit largely renames variables.

Also set lua-resty-jwt's algorithm whitelist to allow
for both, HS256 and RS256 (eventually, only RS256 will
be allowed, but for development as awell as upgrade
purposes the code in auth/open.lua should work for
both key types).