Enabled GC of nested container sandboxes by the Mesos agent.
This flips a flag that instructs the Mesos agent to garbage collect
the sandboxes of nested containers, just like the agent currently
GC's executor sandboxes.  Nested containers always run underneath
another active executor, and have in the past been garbage collected
alongside its parent executor.  However, because the default
executor can now potentially run infinite tasks in its lifetime,
the tasks (which are nested containers) can pile up sandboxes.

Note that this flag is technically a breaking change.  A task can
potentially rely on an artifact inside a sibling task's sandbox.
If the sibling task has exited, the sibling's sandbox can now be
garbage collected.  This scenario is however, considered to be
bad practice, as shared artifacts should be placed in the parent's
sandbox or a persistent volume.