Updated Windows variant of 'clean()' to use utility functions.
Previously, we were using a docker container (which we still do on
Linux) in order to remove these directories. It's not clear why we do
this on Linux (for cross platform compatibility?), but at least on
Windows we know its safe to just use the OS independent utility
functions without requiring a docker container.

These utility functions handle symlinks better, as well as perform
retries when the operating system does not release the files quick
enough for deletion and renaming.