Changed pkgpanda to manage deletion of temp files manually.
This change was motivated by trying to be cross platform compatible with
Windows. Usage of temporary files caused problems on Windows as they
are opened, then reopened down the callstack. In other words, reopening
an already opened file seems to cause problems on Windows.

To address this, temporary files are now created, then passed down,
and then deleted if necessary after they are used.