adminrouter: test harness: adjust to authn arch adjustments
Especially adjust to the fact that we have removed a design
flaw from the system as of which the IAM was in the hot path
for authorization decisions.

Authorization here is done truly via authentication, i.e.
valid authentication proof is sufficient. With the IAM in the
hot path we could also have real server-side sessions (which
we don't want, and this change here simply means that we
consistently use client-side sessions).