Add Mesos patches to ensure TEARDOWN is sent in v1 Java shim.
Destruction of the v1 library used in the v1 Java scheduler shim
is not always under our control. For example, the JVM can call JNI
`finalize()` if the Java scheduler nullifies its reference to the
V1Mesos library immediately after sending `TEARDOWN`. We want to make
sure that `TEARDOWN` is sent before the Mesos library is destructed.