Updated pkgpanda 'test_check.py' to run tests on Windows.
To do this, we copied a '.sh' script from the linux variant under
pkgpanda/test_resources/unix/opt/mesosphere/packages
to
pkgpanda/test_resources/windows/opt/mesosphere/packages
and changed it to a '.ps1' script

We also changes a 'not_executable.py' file to a 'not_supported.dummy'
file for windows.

This simulates what the actual installed package on a DC/OS node will
look like at the time a windows variant of that package is installed on
a windows node.

Since the check code for windows *only* works for '.py' and '.ps1' files
(see code in pkgpanda/cli.py which manually invokes the proper
interpreter for these file types), we will need to keep these tests in
sync with any additional file types allow in the CLI code in the future.