fixes dcos-diagnostics integration tests for ifconfig removal
The assert required a certain index on the ouput of dcos diagnostics
commands (see gen/dcos-config.yaml), which does change as we removed
one command from the middle of the list. Also not all commands were covered,
which made it hard to grep for.

Backports: https://github.com/dcos/dcos/pull/3928