Attempt to reduce flakes and gather data on bundle generation.
This ups the timeout for bundle generation to 20 minutes instead of 2 to
ensure that the cluster has enough time to generate the bundle which
is the cause of the vast majority of the flakes for this test.

It also adds logging the time taken to generate the bundle to the output
to get a better idea of how much time it really takes to generate the
bundle.