cockroach: use exhibitor_wait as ExecStartPre
This fixes DCOS_OSS-4601. Without this step
register.py is highly likely to find the
ZooKeeper cluster in a state where individual
ZooKeeper instances are still in standalone
mode. This on the other hand leads to individual
CockroachDB instances being launched without
being aware of each other. The exhibitor_wait
method is applied by using a noop bootstrapper
for the dcos-cockroach unit.