DCOS_OSS-4531 - Fix flaky test by making it more resilient
This PR fixes the flakyness of the test by making it more resilient to
changing file types. Since we do not care about the file types but only
that the files exist, this change makes sure that the test passes even
if the file we expect exists in a gzipped version.