DCOS-15890 - Check Docker is running before version check
This adds a proper message in case Docker is not running instead of
telling the user that the Docker version doesn't match.

`docker info` sets its exit-code according to whether Docker is running,
so we can use that in a platform-independent way to make sure it is.
See: https://docs.docker.com/config/daemon/#check-whether-docker-is-running

So before we check the Docker version, we check for the exit code of
`docker info`.