DCOS-16098 - Correct filename in dcos_generate_config.ee.sh help
Before this change the ARGV0 that's used by `dcos_install` when printing
the `--help` output was always `dcos_generate_config.sh`, even when it
was built for another variant. In the case of the ticket: `ee`.

This change checks whether `$BOOTSTRAP_VARIANT` is empty or not. And if
it's not empty, it uses it to construct the correct filename.