DCOS_OSS-4613 - Use default `ssh_user` in deprecated CLI installer validations
Requiring the `ssh_user` attribute in `config.yaml` was part of the
CLI installer.

But even though the CLI installer was deprecated, users now run into
validation errors when they run `./dcos_generate_config.sh` without
specifying a `ssh_user` in their `config.yaml`, even though we explictly
say in our documentation that the field is not required.

This is now the first step to fix the ticket and resolve the issue for
the customer: by setting a default value for `ssh_user` users won't run
into a validation error, but if they change the value, we still validate
it.

It's the smallest possible fix.

The proper solution would be to remove the deprecated code and rip out
the the parts of the CLI installer that are not needed anymore (I did a
first spike for this in https://github.com/dcos/dcos/pull/4377 and could
remove ~1000 lines of code without breaking the officially supported
behaviour)

(cherry picked from commit eaff454200578f1bcea243f556af926da75c57fb)