DCOS_OSS-4097 - Explicitly create user groups in pkgpanda
pkgpanda previously relied on the fact that the `useradd` command
creates a matching group for a newly-created user.

But that's not always the case and only worked "by accident" because it
was set in the `USERGROUPS_ENAB` setting in `/etc/login.defs`. (See
https://linux.die.net/man/8/useradd for more information about this)

This commit makes the behaviour explicit and work regardless of
preconfigured settings by using the `--user-group`/`-U` argument to
`useradd` in case no other group was configured with `-g`, because those
are in conflict.