Tune Telegraf config for busy agents
This commit edits Telegraf config to help prevent dropped metrics on
busy agents. This is accomplished by increasing the number of metrics
output pluggins can buffer before they start dropping metrics
(`metric_buffer_limit`), and by increasing the collection `interval` for
the `dcos_statsd` input plugin, which reduces the rate at which it
collects metrics, thus reducing pressure on output buffers.

This commit also reduces the default `interval` for all input plugins to
10s. Most inputs don't need to be throttled on a busy agent, since they
emit a consistent number of metrics that doesn't increase as workloads
are executed on an agent.