Fix fault domain metric tags
The Telegraf startup script intends to run the fault domain detect
script and extract the names of the region and zone from its output, in
order to set these values in environment variables that can be
referenced in Telegraf config. The current implementation is setting
these environment variables in a subshell, which is then immediately
exited, discarding the environment variables, and causing Telegraf to
tag metrics with the names of the env vars it intends to reference.

This commit fixes the startup script by setting these environment
variables in the same shell that execs Telegraf. This is accomplished
by modifying extract_fault_domain.py to print either the zone or region
with each run, and then running it twice to get both the zone and region
names and assign them to env vars.