Changed order of containerizers on Mesos agents.
This is a subtle change that changes the processing order for
operations associated with launching, monitoring, or stopping
containers.  Effectively, the Mesos agent will try the Mesos
containerizer before the Docker containerizer.

This change improves performance of launching Mesos containers when
there are a large number of Docker containers running on the same
agent.  The Docker containerizer tends to spend most of its time
waiting on the Docker daemon to respond, sometimes blocking further
progress if there are too many containers.  Sometimes, this waiting
will cause Mesos containers to take excessively long times to launch,
as we get blocked while executing the launch command on each
containerizer, *in order*.  If the Mesos containerizer comes first,
then we skip the wait entirely.

This is unlikely to cause the opposite scenario (too many Mesos
containers blocking a Docker container from launching) because the
Mesos containerizer does not have long blocking sections.