Made `make_disk_resources.py` report total size instead of free space.
We have seen that when a user removes `mesos-resources` and restarts the
agent with a new ID, the new agent reports less disk space because the
disk size calculation is based on free space before previous sandboxes
has been GC'ed. This patch makes the disk size calculation based on
total size with some headroom instead.