DCOS_OSS-5083 Use PATH for coreutils and procps bins
Rather than supply an explicit path to the binaries provided by the
coreutils or procps packages, fetch them from the PATH. These are
binaries such as `cp` or `kill` and tend to be assumed to be in the
default PATH on any system. However, these are different prefixes on
RHEL-based systems and Debian-based systems. Rather than be explicit on
the prefix, let's assume the PATH includes these items. This will always
be true on Redhat and Ubuntu installations.

Signed-off-by: Chris Gianelloni <cgianelloni@applause.com>